# swagger-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TweetAI API",
    "description": "API documentation for the TweetAI platform.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "Autobot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Autobot ID"
          },
          "name": {
            "type": "string",
            "description": "Autobot name"
          },
          "username": {
            "type": "string",
            "description": "Autobot username"
          },
          "email": {
            "type": "string",
            "description": "Autobot email"
          },
          "phone": {
            "type": "string",
            "description": "Autobot phone number"
          }
        },
        "example": {
          "id": "1",
          "name": "John Doe",
          "username": "johndoe",
          "email": "johndoe@example.com",
          "phone": "08012345678"
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Post ID"
          },
          "title": {
            "type": "string",
            "description": "Post title"
          },
          "body": {
            "type": "string",
            "description": "Post body"
          },
          "autobotId": {
            "type": "string",
            "description": "ID of the Autobot who created the post"
          }
        },
        "example": {
          "id": "1",
          "title": "Autobot Post",
          "body": "This is a sample post created by an Autobot.",
          "autobotId": "1"
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Comment ID"
          },
          "body": {
            "type": "string",
            "description": "Comment text"
          },
          "postId": {
            "type": "string",
            "description": "ID of the post this comment belongs to"
          }
        },
        "example": {
          "id": "1",
          "body": "This is a sample comment.",
          "postId": "1"
        }
      }
    }
  },
  "paths": {
    "/get-all-autobots": {
      "get": {
        "summary": "Get all Autobots",
        "tags": [
          "Endpoints"
        ],
        "responses": {
          "200": {
            "description": "List of all Autobots.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "example": {
                      "id": 1,
                      "name": "Optimus Prime",
                      "username": "Optimus",
                      "email": "optimus@gmail.com",
                      "phone": 8012345678
                    },
                    "createdAt": "2021-08-02T14:00:00.000Z",
                    "updatedAt": "2021-08-02T14:00:00.000Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/get-count": {
      "get": {
        "summary": "Get total count of Autobots",
        "tags": [
          "Endpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Count of Autobots.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "example": {
                    "count": 507
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/get-post-comments/{postId}": {
      "get": {
        "summary": "Get comments of a post",
        "tags": [
          "Endpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments for a specific post.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "example": null,
                  "id": 1,
                  "body": "Comment Body",
                  "PostId": 1,
                  "createdAt": "2021-08-02T14:00:00.000Z",
                  "updatedAt": "2021-08-02T14:00:00.000Z"
                }
              }
            }
          },
          "404": {
            "description": "No Comment found."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/get-bot-posts/{id}": {
      "get": {
        "summary": "Get Autobot by ID",
        "tags": [
          "Endpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Autobot ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of posts for a specific Autobot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Autobot"
                }
              }
            }
          },
          "404": {
            "description": "Autobot not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    }
  },
  "tags": []
}